requirements.txt
Django>=4.1.10
djangorestframework>=3.13.1
psycopg2>=2.9.3,<2.10
drf-spectacular>=0.22.1
django-allauth>=0.61.1
#Pillow>=9.1.0,<9.2
#uwsgi>=2.0.20,<2.1